requirements.txt
customtkinter
certifi
portablemc
CTkMessagebox
CTkListbox
beautifulsoup4
requests
pyyaml